requirements.txt
dagster==1.11.15
google-cloud-bigquery>=3.38.0
google-search-results>=2.4.2
dagster-webserver==1.11.15
